# src/tsoa/swagger.json
{
	"components": {
...
		"schemas": {
...
			"FailureResponse": {
				"properties": {
					"error": {
						"type": "string"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Globals.PlanetType": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"Planet": {
				"properties": {
					"planetID": {
						"type": "number",
						"format": "double"
					},
					"ownerID": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"posGalaxy": {
						"type": "number",
						"format": "double"
					},
					"posSystem": {
						"type": "number",
						"format": "double"
					},
					"posPlanet": {
						"type": "number",
						"format": "double"
					},
					"lastUpdate": {
						"type": "number",
						"format": "double"
					},
					"planetType": {
						"$ref": "#/components/schemas/Globals.PlanetType"
					},
					"image": {
						"type": "string"
					},
					"diameter": {
						"type": "number",
						"format": "double"
					},
					"fieldsCurrent": {
						"type": "number",
						"format": "double"
					},
					"fieldsMax": {
						"type": "number",
						"format": "double"
					},
					"tempMin": {
						"type": "number",
						"format": "double"
					},
					"tempMax": {
						"type": "number",
						"format": "double"
					},
					"metal": {
						"type": "number",
						"format": "double"
					},
					"crystal": {
						"type": "number",
						"format": "double"
					},
					"deuterium": {
						"type": "number",
						"format": "double"
					},
					"energyUsed": {
						"type": "number",
						"format": "double"
					},
					"energyMax": {
						"type": "number",
						"format": "double"
					},
					"metalMinePercent": {
						"type": "number",
						"format": "double"
					},
					"crystalMinePercent": {
						"type": "number",
						"format": "double"
					},
					"deuteriumSynthesizerPercent": {
						"type": "number",
						"format": "double"
					},
					"solarPlantPercent": {
						"type": "number",
						"format": "double"
					},
					"fusionReactorPercent": {
						"type": "number",
						"format": "double"
					},
					"solarSatellitePercent": {
						"type": "number",
						"format": "double"
					},
					"bBuildingId": {
						"type": "number",
						"format": "double"
					},
					"bBuildingEndTime": {
						"type": "number",
						"format": "double"
					},
					"bBuildingDemolition": {
						"type": "boolean"
					},
					"bHangarQueue": {
						"type": "string"
					},
					"bHangarStartTime": {
						"type": "number",
						"format": "double"
					},
					"bHangarPlus": {
						"type": "boolean"
					},
					"destroyed": {
						"type": "boolean"
					}
				},
				"required": [
					"planetID",
					"ownerID",
					"name",
					"posGalaxy",
					"posSystem",
					"posPlanet",
					"lastUpdate",
					"planetType",
					"image",
					"diameter",
					"fieldsCurrent",
					"fieldsMax",
					"tempMin",
					"tempMax",
					"metal",
					"crystal",
					"deuterium",
					"energyUsed",
					"energyMax",
					"metalMinePercent",
					"crystalMinePercent",
					"deuteriumSynthesizerPercent",
					"solarPlantPercent",
					"fusionReactorPercent",
					"solarSatellitePercent",
					"bBuildingId",
					"bBuildingEndTime",
					"bBuildingDemolition",
					"bHangarQueue",
					"bHangarStartTime",
					"bHangarPlus",
					"destroyed"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"BuildOrderItem": {
				"properties": {
					"unitID": {
						"type": "number",
						"format": "double"
					},
					"amount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"unitID",
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Event": {
				"properties": {
					"eventID": {
						"type": "number",
						"format": "double"
					},
					"ownerID": {
						"type": "number",
						"format": "double"
					},
					"mission": {
						"type": "number",
						"format": "double"
					},
					"fleetlist": {
						"type": "string"
					},
					"startID": {
						"type": "number",
						"format": "double"
					},
					"startType": {
						"$ref": "#/components/schemas/Globals.PlanetType"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"endID": {
						"type": "number",
						"format": "double"
					},
					"endType": {
						"$ref": "#/components/schemas/Globals.PlanetType"
					},
					"endTime": {
						"type": "number",
						"format": "double"
					},
					"loadedMetal": {
						"type": "number",
						"format": "double"
					},
					"loadedCrystal": {
						"type": "number",
						"format": "double"
					},
					"loadedDeuterium": {
						"type": "number",
						"format": "double"
					},
					"returning": {
						"type": "boolean"
					},
					"inQueue": {
						"type": "boolean"
					},
					"processed": {
						"type": "boolean"
					}
				},
				"required": [
					"eventID",
					"ownerID",
					"mission",
					"fleetlist",
					"startID",
					"startType",
					"startTime",
					"endID",
					"endType",
					"endTime",
					"loadedMetal",
					"loadedCrystal",
					"loadedDeuterium",
					"returning",
					"inQueue",
					"processed"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DestroyPlanetRequest": {
				"properties": {
					"planetID": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"planetID"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RenamePlanetRequest": {
				"properties": {
					"planetID": {
						"type": "number",
						"format": "double"
					},
					"newName": {
						"type": "string"
					}
				},
				"required": [
					"planetID",
					"newName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Ships": {
				"properties": {
					"planetID": {
						"type": "number",
						"format": "double"
					},
					"smallCargoShip": {
						"type": "number",
						"format": "double"
					},
					"largeCargoShip": {
						"type": "number",
						"format": "double"
					},
					"lightFighter": {
						"type": "number",
						"format": "double"
					},
					"heavyFighter": {
						"type": "number",
						"format": "double"
					},
					"cruiser": {
						"type": "number",
						"format": "double"
					},
					"battleship": {
						"type": "number",
						"format": "double"
					},
					"colonyShip": {
						"type": "number",
						"format": "double"
					},
					"recycler": {
						"type": "number",
						"format": "double"
					},
					"espionageProbe": {
						"type": "number",
						"format": "double"
					},
					"bomber": {
						"type": "number",
						"format": "double"
					},
					"solarSatellite": {
						"type": "number",
						"format": "double"
					},
					"destroyer": {
						"type": "number",
						"format": "double"
					},
					"battlecruiser": {
						"type": "number",
						"format": "double"
					},
					"deathstar": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"planetID",
					"smallCargoShip",
					"largeCargoShip",
					"lightFighter",
					"heavyFighter",
					"cruiser",
					"battleship",
					"colonyShip",
					"recycler",
					"espionageProbe",
					"bomber",
					"solarSatellite",
					"destroyer",
					"battlecruiser",
					"deathstar"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BuildShipsRequest": {
				"properties": {
					"planetID": {
						"type": "number",
						"format": "double"
					},
					"buildOrder": {
						"items": {
							"$ref": "#/components/schemas/BuildOrderItem"
						},
						"type": "array"
					}
				},
				"required": [
					"planetID",
					"buildOrder"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "access-token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "ugamela api",
		"version": "1.0.0",
		"description": "bla",
		"license": {
			"name": "AGPL-3.0"
		},
		"contact": {
			"email": "api@ugamela.org",
			"name": "the Administrator"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/planets/planetList": {
			"get": {
				"operationId": "GetAllPlanets",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Planet"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Planets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/planets/planetList/{userID}": {
			"get": {
				"operationId": "GetAllPlanetsOfUser",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Planet"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Planets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/planets/movement/{planetID}": {
			"get": {
				"operationId": "GetMovement",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Event"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Planets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "planetID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/planets/destroy": {
			"post": {
				"operationId": "DestroyPlanet",
				"responses": {
					"200": {
						"description": ""
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Planets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DestroyPlanetRequest"
							}
						}
					}
				}
			}
		},
		"/planets/rename": {
			"post": {
				"operationId": "RenamePlanet",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Planet"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Planets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RenamePlanetRequest"
							}
						}
					}
				}
			}
		},
		"/planets/{planetID}": {
			"get": {
				"operationId": "GetPlanetByID",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Planet"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Planets"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "planetID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/ships/{planetID}": {
			"get": {
				"operationId": "GetAllShipsOnPlanet",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Ships"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Ships"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "planetID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/ships/build": {
			"post": {
				"operationId": "BuildShips",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Planet"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Ships"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BuildShipsRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://127.0.0.1:3000/v1"
		}
	]
}